+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Add rule type 'OS' to `New-GRRHunt`. Use os_windows, os_linux or os_darwin
     as filter ([#3](https://github.com/swisscom/PowerGRR/issues/3)).
 * Add new cmdlet for getting a list of all available flows
-    (`Get-GRRFlowDescriptor`) and add corresponding Pester tests.
+    (`Get-GRRFlowDescriptor`) and add corresponding Pester tests ([#4](https://github.com/swisscom/PowerGRR/issues/4)).
+* Add flow type **ArtifactCollectorFlow** in `Invoke-GRRFlow` and `New-GRRHunt` ([#2](https://github.com/swisscom/PowerGRR/issues/2))
 
 ### Changed